+++ packages/nap/src/outbox/index.ts
@@ -22,7 +22,7 @@
 export { DOMAIN } from './types.js';
 
 export type {
-  OutboxStrategy,
+  RelayEventResult,
   OutboxEventOptions,
   OutboxQueryOptions,
   OutboxSubscribeOptions,
@@ -40,7 +40,6 @@ export type {
   OutboxQueryResultMessage,
   OutboxSubscribeMessage,
   OutboxEventMessage,
-  OutboxEoseMessage,
   OutboxClosedMessage,
   OutboxCloseMessage,
   OutboxPublishMessage,