+++ apps/conformance/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @napplet/conformance-web
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [6ccb056]
+  - @napplet/conformance@0.12.0
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [284e100]
+  - @napplet/conformance@0.11.0
+
 ## 0.0.11
 
 ### Patch Changes